views cleanup, event app to view package
