+++ losuj_to/events/urls.py
@@ -1,21 +1,27 @@
 from django.urls import path
-from events.views import (
-    BulkUserRegistration,
-    EmailSendInvitationsWait,
-    EmailSendInvitationsWaitStream,
-    EventActivate,
+from events.views.crud import (
+    EventActivateView,
     EventCreateView,
-    EventDeactivate,
-    EventExcludesCreate,
-    EventExcludesUpdate,
-    EventParticipantsCreateView,
-    EventParticipantsUpdateView,
-    EventSendInvitations,
-    EventSendReminderSingle,
-    EventSummarry,
+    EventDeactivateView,
     EventUpdateView,
-    ListOfEvents,
-    ParticipantEventView,
+)
+from events.views.notifications import (
+    InvitationReminderView,
+    InvitationSendView,
+    InvitationStreamWaitView,
+    InvitationWaitView,
+)
+from events.views.participants import (
+    BulkUserRegistration,
+    ParticipantCreateView,
+    ParticipantExcludeCreateView,
+    ParticipantExcludeUpdateView,
+    ParticipantUpdateView,
+)
+from events.views.summaries import (
+    EventAdminDetailView,
+    EventListView,
+    EventUserDetailView,
 )
 
 urlpatterns = [
@@ -23,7 +29,7 @@ urlpatterns = [
     path("event_update/<int:pk>", EventUpdateView.as_view(), name="event_update"),
     path(
         "event_create_participants/",
-        EventParticipantsCreateView.as_view(),
+        ParticipantCreateView.as_view(),
         name="event_participants",
     ),
     path(
@@ -31,57 +37,61 @@ urlpatterns = [
         BulkUserRegistration.as_view(),
         name="bulk_registration",
     ),
-    path("event_excludes/", EventExcludesCreate.as_view(), name="event_excludes"),
-    path("event_summary/<int:pk>", EventSummarry.as_view(), name="event_summary"),
-    path("event_view/<int:pk>", ParticipantEventView.as_view(), name="event_view"),
+    path(
+        "event_excludes/", ParticipantExcludeCreateView.as_view(), name="event_excludes"
+    ),
+    path(
+        "event_summary/<int:pk>", EventAdminDetailView.as_view(), name="event_summary"
+    ),
+    path("event_view/<int:pk>", EventUserDetailView.as_view(), name="event_view"),
     path(
         "event_view/<slug:hash>",
-        ParticipantEventView.as_view(),
+        EventUserDetailView.as_view(),
         name="event_view_by_hash",
     ),
     path(
         "event_update_participants/<int:pk>",
-        EventParticipantsUpdateView.as_view(),
+        ParticipantUpdateView.as_view(),
         name="event_participants_update",
     ),
     path(
         "event_update_excludes/<int:pk>",
-        EventExcludesUpdate.as_view(),
+        ParticipantExcludeUpdateView.as_view(),
         name="event_excludes_update",
     ),
     path(
         "event_toggle_active/<int:pk>",
-        EventActivate.as_view(),
+        EventActivateView.as_view(),
         name="event_toggle_active",
     ),
     path(
         "event_deactivate/<int:pk>",
-        EventDeactivate.as_view(),
+        EventDeactivateView.as_view(),
         name="event_deactivate",
     ),
     path(
         "send_invitations/<int:pk>",
-        EventSendInvitations.as_view(),
+        InvitationSendView.as_view(),
         name="send_invitations",
     ),
     path(
         "send_reminder/<int:pk>/<int:participant_id>",
-        EventSendReminderSingle.as_view(),
+        InvitationReminderView.as_view(),
         name="send_reminder",
     ),
     path(
         "send_invitations_wait/<int:pk>/<str:task_ids>",
-        EmailSendInvitationsWait.as_view(),
+        InvitationWaitView.as_view(),
         name="send_invitations_wait",
     ),
     path(
         "event_list/",
-        ListOfEvents.as_view(),
+        EventListView.as_view(),
         name="event_list",
     ),
     path(
         "status_stream/<int:pk>/<str:task_ids>",
-        EmailSendInvitationsWaitStream.as_view(),
+        InvitationStreamWaitView.as_view(),
         name="event_send_invitation_status_stream",
     ),
 ]
