+++ losuj_to/events/views/home.py
@@ -1,0 +1,5 @@
+from django.views.generic import TemplateView
+
+
+class HomeView(TemplateView):
+    template_name = "users/home.html"
